mash-light: Add virtuals to generate the shader and update uniforms
There is now a virtual method used to generate the shader snippet for
each light. The idea is that the light box will call this for each
light before painting whenever a light is added or removed. The
virtual takes two GStrings - one for declaring uniforms or global
attributes and one for the code in the main function. Each light
subclass should append to these strings with its own lighting
calculations.

There is also a virtual method which gets called before painting any
other children of a light box which the light can use to update its
uniforms. It is expected that the program will be in use during this
function so the method implementation can just directly call
cogl_program_uniform_*.

The class default methods add the uniforms needed to handle the three
light colors that all lights have. No code is generate in the main
function.