MashPointLight: Add the attenuation properties
A point light has three properties to make a formula for the
attenuation: constant, linear and quadratic. Constant just adds a
number to the attenuation value, linear is multiplied by the distance
from the light to the vertex and quadratic is multiplied by the square
of the distance. The results of all of these are added together to get
the final value.