Implement directional lights
This adds the shader code for directional lights. This is mostly a
strimmed down version of the code for point lights except the code for
setting the direction uniform is a bit more complicated because it
needs to use the CPU equivalent of the gl_NormalMatrix to safely
transform the direction vector.