Add an example program with controls to manipulate lights
This adds an example program that is only built if Mx is detected in
the configure script. It is similar to the viewer except that it adds
three lights to the scene. There are controls to manipulate every
property of the material and the lights. Each light is a different
type. Initially the lights are all set to emit zero light for each
light effect which effectively disables them. This makes it easier to
test the look of an individual light type by enabling just one light.