Define a version numbering system and use it
The configure.ac script now contains a short description of the
version numbering system used for Mash and resets the version number
down to 0.0.1. The major and minor versions now form the API
version. This gets concatenated into the library name so that
incompatible versions of Mash can be parallel installed.

The libtool interface version is managed separately from the package
version number as suggested by the libtool manual.