mash-light-box: Copy the texture coordinates in the vertex shader
The texture coordinate attribute from the vertex shader needs to be
copied to the varying output or texturing isn't going to work. This
isn't ideal because it's limited to using a CoglMaterial that only has
one layer. Hopefully this can be fixed when Cogl gets support for a
shader snippet API.