Rename MashLightBox to MashLightSet and make it no longer an actor
The lights no longer have to be in the same container as the models
for lighting to work. Instead, the lights can be anywhere in the actor
hierarchy. They are grouped together in a MashLightSet which is
separate from the actor's parent. MashLightSet is now just a linked
list of lights. MashModel is now aware of the lighting and it has a
property to set the MashLightSet to use. Before painting the mash
model, it will query the MashLightSet for the current program and set
it on its material before painting. This avoids the need to use
cogl_program_use which is deprecated in Clutter 1.6 because it ends up
creating a one-shot pipeline for every primitive.

The 'update_uniforms' method of MashLight is no longer called during
the paint sequence of the light because there is no longer any way to
guarantee that this will be called before the first model is
painted. Instead the method is called the first time a model tries to
paint with the light set every frame.