Use the cogl_* names for the builtin GLSL variables
The GLSL code for the lighting now uses the Cogl names for the builtin
variables instead of the gl_* namespaced variables. In theory these
should be portable to GLES2. There is no Cogl builtin for the normal
matrix so Mash now calculates and uploads its own matrix. There is
also no builtin for the lighting material parameters so Mash defines
its own uniform and extracts these from the material.